// generated/src/app/api-models/stage-auth-advisor.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthAdvisor {
  batchId: string;
  transactionId: number;
  seqNo: number;
  provId: string;
  contactDate: Date;
  recommCode: string;
  decisionDate: Date;
  decision: string;
  service: string;
  comments: string;
  insertDatetime: Date;
}